Ben Stephenson

+++ Ch-3 Loop Exercises/Ex-77.py
@@ -1,8 +1,7 @@
 import math
-import math
 
 a = input("Enter the number in binary: ")
 num = 0
 for i in range(len(a)):
-    num += int(a[-i-1]) * math.pow(2, i)
+    num += int(a[-i-1]) * int(math.pow(2, i))
 print(num)
